Update Players Asynchronously
Rather than just sending packets straight to the
other clients, we will batch updates. This lets
us save on bandwidth and makes room for
spatial partitioning in the future.

+++ fom-network/include/fom-network/packets/WorldUpdate.h
@@ -12,7 +12,7 @@ constexpr int MAX_WORLD_UPDATES = 100;
 struct WorldUpdate {
   uint32_t playerId;
   uint32_t unknown1;
-  uint16_t updateCount;
+  uint8_t updateCount;
   Type::WorldUpdate updates[MAX_WORLD_UPDATES];
 };
 #pragma pack(pop)
